requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
httpx
pydantic